feat: remove duckdbio.ingest()
It's a legacy function that has been replaced by the new API package.

Fixes SparkBeyond/aoa#388
